# docs/openapi.json
{
...
  "openapi": "3.1.0",
  "info": {
    "title": "PitchFlow API",
    "description": "API documentation for PitchFlow - AI-assisted sponsorship proposal workspace",
    "version": "1.0.0",
    "contact": {
      "name": "Rectoverso Media",
      "email": "support@rectoverso.id"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://pitchflow.vercel.app",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/proposals": {
      "get": {
        "tags": ["Proposals"],
        "summary": "Get all proposals",
        "description": "Retrieve all proposals for the authenticated user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of proposals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Proposal" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Proposals"],
        "summary": "Create a new proposal",
        "description": "Create a new proposal for the authenticated user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "brand_name", "pic_sales"],
                "properties": {
                  "title": { "type": "string", "maxLength": 200 },
                  "brand_name": { "type": "string", "maxLength": 200 },
                  "pic_sales": { "type": "string", "maxLength": 100 },
                  "program": { "type": "string", "maxLength": 200 },
                  "industry": { "type": "string", "maxLength": 100 },
                  "sponsorship_type": { "type": "string", "maxLength": 100 },
                  "year": { "type": "integer", "minimum": 2020, "maximum": 2100 },
                  "status": {
                    "type": "string",
                    "enum": ["new_brief", "drafting", "need_input", "revised", "ready"]
                  },
                  "result": {
                    "type": "string",
                    "enum": ["won", "pitched", "lost", "template"]
                  },
                  "brief_id": { "type": "string", "format": "uuid" },
                  "deadline": { "type": "string", "maxLength": 50 },
                  "slides_count": { "type": "integer", "minimum": 0 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proposal created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Proposal" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/sales-comments": {
      "get": {
        "tags": ["Sales"],
        "summary": "Get sales comments",
        "description": "Retrieve comments for a proposal",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "proposal_id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of comments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/SalesComment" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Sales"],
        "summary": "Create a comment",
        "description": "Add a new comment to a proposal",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["proposal_id", "content"],
                "properties": {
                  "proposal_id": { "type": "string", "format": "uuid" },
                  "content": { "type": "string", "maxLength": 2000 },
                  "parent_id": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/SalesComment" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Proposal": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "brand_name": { "type": "string" },
          "program": { "type": "string" },
          "industry": { "type": "string" },
          "sponsorship_type": { "type": "string" },
          "year": { "type": "integer" },
          "status": {
            "type": "string",
            "enum": ["new_brief", "drafting", "need_input", "revised", "ready"]
          },
          "result": {
            "type": "string",
            "enum": ["won", "pitched", "lost", "template"]
          },
          "pic_sales": { "type": "string" },
          "deadline": { "type": "string" },
          "last_activity": { "type": "string", "format": "date-time" },
          "slides_count": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "created_by": { "type": "string", "format": "uuid" }
        }
      },
      "SalesComment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "proposal_id": { "type": "string", "format": "uuid" },
          "user_id": { "type": "string", "format": "uuid" },
          "user_name": { "type": "string" },
          "user_role": { "type": "string", "enum": ["Sales", "ACS", "Supervisor"] },
          "content": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "replies": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SalesComment" }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request - validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": {
              "success": false,
              "error": "Validation failed: brand_name is required"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - authentication required",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": {
              "success": false,
              "error": "Authentication required"
            }
          }
        }
      },
...
    }
  }
}